+++ MOP/src/GameObjects/Items/ItemHook.cs
@@ -158,8 +158,8 @@ namespace MOP
             // If Bottle Recycling mod is present, prevent destroying beer bottles
             if (CompatibilityManager.instance.BottleRecycling)
             {
-                ModConsole.Print("[MOP] Beer bottles won't be destroyed, because Bottle Recycling mod is present. " +
-                    "Disable 'Destroy empty beer bottles' in the MOP settings.");
+                ModConsole.Print("<color=yellow>[MOP] Beer bottles won't be destroyed, because Bottle Recycling mod is present. " +
+                    "Disable 'Destroy empty beer bottles' in the MOP settings.</color>");
                 return;
             }