+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 ### Added
 
+- Added support for rules!
+  - From now on, all mod compatibility is done via the remote server
+  - No need to update MOP in order to add mod compatibility! (At least in most cases)
+  - Mod compatibility rules can be updated if needed! MOP checks for rules update every week (you can force new check in the settings)
 - Engine renderers are now disabled, if player is in the car, and hood is attached
+- Added disclaimer for pirated game copies
 
 ## 2.4.3 (19.03.2020)